Found mistake (not fixed yet)
The way alpha-beta is implemented now is the way we have to work with it. But this only works when we are going deeper, so more often are in isMaximisingPlayer during the recursive loops.

Only then alpha-beta has it benefits over normal minimax.

(now it only says no move proposed when 'if beta <= alpha': then with break it breaks the for loop and we never come to self.propose loop).

so TWO OPTIONS:
1. go deeper, more often be in isMaximisingPlayer in recursive loop
2. make a special a-b variant for the current minimax

I think 1 is the way to go.